### 2024-01-25 #6282 @JBorrow
Hi @tberlok! I will be the handling editor for this submission. You're right, this looks like a very nice package that has lots of similarities with swiftsimio.

The next step is for us to find some reviewers for the submission. If you have any suggestions, I invite you to let me know by putting their GitHub usernames in a comment, but please do not 'mention' them at this stage (e.g. you would write JBorrow, not @JBorrow).
