### 2024-01-27 #6282 @kyleaoman
> Hi @kyleaoman, would you be interested in reviewing this package for JOSS? Our review guidelines are available here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html should you want to take a look before agreeing to review.

Yes, I can review.
