### 2024-01-29 #6282 @JBorrow
Apologies for the typos there, but thank you so much @kyleaoman and @ttricco for agreeing to review! We will move forward with the review thread next. As a reminder our reviewer guidelines are available here: https://joss.readthedocs.io/en/latest/reviewer_guidelines.html.

JOSS is a little different to typical journals as the review proceeds entirely in public, and uses a checklist-based system mainly based upon the software package and documentation. Please do let me know if you have any questions.
