### 2024-01-29 #6296 @editorialbot
[REVIEW]: Paicos: A Python package for analysis of (cosmological) simulations performed with Arepo
**Submitting author:** @tberlok (<a href="http://orcid.org/0000-0003-0466-603X">Thomas Berlok</a>)
**Repository:** https://github.com/tberlok/paicos
**Branch with paper.md** (empty if default branch): paper
**Version:** v0.1.14
