### 2024-01-29 #6296 @JBorrow
Hi @tberlok, it looks like there is still an issue with a missing DOI for the paper. Also, it may be worth me mentioning that the latest information on how to cite the SWIFT code is available here: https://swift.strw.leidenuniv.nl/docs/CitingSWIFT/index.html.
