### 2024-01-29 3278411
Fix SWIFT citation and missing DOI
