+++ paper/paper.bib
@@ -22,7 +22,24 @@
   version      = {v1.0.0},
   doi          = {10.5281/zenodo.21703},
   url          = {https://doi.org/10.5281/zenodo.21703}
-} 
+}
+
+@ARTICLE{Schaller2023,
+  author = {{Schaller}, Matthieu and others},
+  title = "{Swift: A modern highly-parallel gravity and smoothed particle hydrodynamics solver for astrophysical and cosmological applications}",
+  journal = {arXiv e-prints},
+  keywords = {Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Cosmology and Nongalactic Astrophysics, Astrophysics - Earth and Planetary Astrophysics, Astrophysics - Astrophysics of Galaxies, Computer Science - Distributed, Parallel, and Cluster Computing},
+  year = 2023,
+  month = may,
+  eid = {arXiv:2305.13380},
+  pages = {arXiv:2305.13380},
+  doi = {10.48550/arXiv.2305.13380},
+  archivePrefix = {arXiv},
+  eprint = {2305.13380},
+  primaryClass = {astro-ph.IM},
+  adsurl = {https://ui.adsabs.harvard.edu/abs/2023arXiv230513380S},
+  adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
 
 @ARTICLE{yt-Turk,
    author = {{Turk}, M.~J. and {Smith}, B.~D. and {Oishi}, J.~S. and {Skory}, S. and
@@ -409,16 +426,17 @@ archivePrefix = {arXiv},
 @ARTICLE{Berlok2022,
        author = {{Berlok}, Thomas},
         title = "{Hydromagnetic waves in an expanding universe - cosmological MHD code tests using analytic solutions}",
-      journal = {arXiv e-prints},
-     keywords = {Astrophysics - Cosmology and Nongalactic Astrophysics},
+      journal = {\mnras},
+     keywords = {hydrodynamics, MHD, waves, cosmology: theory, software: simulations, software: development, Astrophysics - Cosmology and Nongalactic Astrophysics},
          year = 2022,
-        month = mar,
-          eid = {arXiv:2203.11887},
-        pages = {arXiv:2203.11887},
+        month = sep,
+       volume = {515},
+       number = {3},
+        pages = {3492-3511},
+          doi = {10.1093/mnras/stac1882},
 archivePrefix = {arXiv},
        eprint = {2203.11887},
  primaryClass = {astro-ph.CO},
-       adsurl = {https://ui.adsabs.harvard.edu/abs/2022arXiv220311887B},
+       adsurl = {https://ui.adsabs.harvard.edu/abs/2022MNRAS.515.3492B},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
-
