+++ .github/SUPPORT.md
@@ -1,0 +1,4 @@
+Paicos uses GitHub Issues to track bugs and feature requests. Please search the existing issues before filing new issues to avoid duplicates. For new issues, file your bug or feature request as a new Issue.
+
+For instructions on how to use Paicos, please consult the documentation on [readthedocs](https://paicos.readthedocs.io). If you have questions not answered there, then please
+feel free to send an email to tberlok ad nbi.ku.dk.
