### 2024-02-12 69a053d
Create CONTRIBUTING.md
