+++ .github/CONTRIBUTING.md
@@ -1,0 +1,5 @@
+Contributions to Paicos are very welcome! In order to avoid potentially duplicate work, we strongly encourage you to create an issue describing the feature that you intend to develop
+before you start working on it. You are also very welcome to send an email directly to Thomas Berlok at tberlok ad nbi.ku.dk.
+
+Pull requests are required to pass all the automated tests. For new functionality, they are also expected to come with a Python script or notebook example
+illustrating the new funcitonality that has been developed.
