+++ .github/CONTRIBUTING.md
@@ -2,4 +2,4 @@ Contributions to Paicos are very welcome! In order to avoid potentially duplicat
 before you start working on it. You are also very welcome to send an email directly to Thomas Berlok at tberlok ad nbi.ku.dk.
 
 Pull requests are required to pass all the automated tests. For new functionality, they are also expected to come with a Python script or notebook example
-illustrating the new funcitonality that has been developed.
+illustrating the new functionality that has been developed.
