### 2024-02-19 tberlok/paicos#47 @kyleaoman
Further detail in contribution guidelines
Suggest fleshing out contribution guidelines a bit - what is the recommended way to install the development version of the package? What is the recommended way to run the test suite locally? Some of this is mentioned in the documentation, linking to those pages as appropriate would make sense.

It would also be useful to include information about expectations from contributors regarding documentations, test coverage, and similar.
