### 2024-02-19 tberlok/paicos#48 @kyleaoman
Dependencies
Will add any comments on dependencies here as I come across them.

To start: the notebook examples depend on matplotlib, which is not a dependency of the package (could be added as an optional dependency) or mentioned in the docs. While most users probably have this installed, I started from a blank virtual environment to catch these sorts of corner cases...
