### 2024-02-19 tberlok/paicos#50 @kyleaoman
Tests requiring gpu should xfail if gpu is not available
It would make sense for tests needing the gpu to be skipped or xfail if a gpu is not available.
