### 2024-02-19 tberlok/paicos#51 @kyleaoman
Typo or error in pytest markers?
Pytest reports the following warning:

```
tests/comoving/test_slicer_writer_and_reader.py:4
  /home/koman/code/paicos/tests/comoving/test_slicer_writer_and_reader.py:4: PytestUnknownMarkWarning: Unknown pytest.mark.order - is this a typo?  You can register custom marks to avoid this warning - for details, see https://docs.pytest.org/en/stable/how-to/mark.html
    @pytest.mark.order(1)
```
