### 2024-02-19 tberlok/paicos#52 @kyleaoman
Typos
May as well gather these together as I notice them during review.

1. "abble" in place of "able" on https://paicos.readthedocs.io/en/latest/configuration.html
2. "get_depencies" should be "get_dependencies" (?) on https://paicos.readthedocs.io/en/latest/configuration.html
3. https://paicos.readthedocs.io/en/latest/examples.html refers to subdirectory "ipython-notebooks" but seems to actually be named "notebook-tutorials".
4. https://paicos.readthedocs.io/en/latest/examples.html refers to "notebook 1a" but notebooks are not numbered (seems like "paicos_basics.ipynb" is probably the one to start with?). In general would be useful to synchronize the labels in the docs (https://paicos.readthedocs.io/en/latest/notebooks/creating_slices_projections.html) and the notebook filenames.
5. In notebook "paicos_basics" sentence missing period "Paicos gives information about what is happening under the hood"
6. Something strange going on in the parameters and return values sections of PaicosQuantity documentation at https://paicos.readthedocs.io/en/latest/autodoc.html?
7. Backslashes behaving weirdly in documentation on readthedocs, e.g. in `PaicosQuantity`:`label` we have "ho or abla imes ec{v}." on https://paicos.readthedocs.io/en/latest/autodoc.html, I guess that's "rho", "nabla" etc.
8. Error in docstring of `paicos.image_creators.nested_projector.NestedProjector` leads to incorrectly formatted parameter list on https://paicos.readthedocs.io/en/latest/autodoc.html
9. Also in docstring of `paicos.image_creators.slicer.Slicer`:`slice_variable`
10. Also in 'Returns' section of `project_variable` from `paicos.image_creators.tree_projector.TreeProjector`
