### 2024-02-19 tberlok/paicos#53 @kyleaoman
API documentation incomplete & bugs
The API documentation section has some incomplete docstrings particularly in the `paicos.derived_variables` section. There are also a number of errors/glitches in the html rendering on readthedocs, some examples are listed in #52 but please check carefully for more.
