### 2024-02-19 tberlok/paicos#54 @kyleaoman
License is not an OSI approved license?
JOSS review checklist asks for an OSI-approved license in the license file. Perhaps this is one, but I don't see the name of the license in the license file so it's difficult to tell.
