### 2024-02-19 #6296 @kyleaoman
I've completed most of my review and have left a number of github issues open for the authors. My checklist above reflects which areas I feel need a little bit of work (but overall the package is in a good state). The exception is the "Performance" item - I should still try to check the functionality of the GPU-accelerated features. My laptop doesn't have a suitable GPU, but I should be able to test this on our local compute cluster. Will try to do this asap.
