### 2024-02-20 tberlok/paicos#55 @kyleaoman
Deprecation warnings in `make checks`
Some deprecations coming from numpy appear in the output of `make checks` after installation.

```
> make checks
make cleanup
make[1]: Entering directory `/cosma/home/durham/dc-oman1/code/paicos'
rm -rf data/test_data
rm -f data/very_small_snap_247.hdf5
rm -f data/reduced_snap2_247.hdf5
make[1]: Leaving directory `/cosma/home/durham/dc-oman1/code/paicos'
cd tests; pytest comoving
rootdir: /cosma/home/durham/dc-oman1/code/paicos
plugins: order-1.2.0
collected 9 items

comoving/test_slicer_writer_and_reader.py .                              [ 11%]
comoving/test_2D_histogram_openmp.py .                                   [ 22%]
comoving/test_arepo_image.py .                                           [ 33%]
comoving/test_binary_tree.py .                                           [ 44%]
comoving/test_compare_projector_with_snap_util.py .                      [ 55%]
comoving/test_compare_slicer_with_snap_util.py .                         [ 66%]
comoving/test_nested_plot_few_particles.py .                             [ 77%]
comoving/test_projector.py .                                             [ 88%]
comoving/test_radial_selection_function.py .                             [100%]

=============================== warnings summary ===============================
tests/comoving/test_slicer_writer_and_reader.py::test_slicer_writer_and_reader
tests/comoving/test_slicer_writer_and_reader.py::test_slicer_writer_and_reader
  /cosma/home/durham/dc-oman1/.virtualenv/paicos_review/lib/python3.10/site-packages/cupy/__init__.py:103: DeprecationWarning: `np.bool8` is a deprecated alias for `np.bool_`.  (Deprecated NumPy 1.24)
    from numpy import bool8  # NOQA

tests/comoving/test_slicer_writer_and_reader.py::test_slicer_writer_and_reader
tests/comoving/test_slicer_writer_and_reader.py::test_slicer_writer_and_reader
  /cosma/home/durham/dc-oman1/.virtualenv/paicos_review/lib/python3.10/site-packages/cupy/__init__.py:114: DeprecationWarning: `np.int0` is a deprecated alias for `np.intp`.  (Deprecated NumPy 1.24)
    from numpy import int0  # NOQA

tests/comoving/test_slicer_writer_and_reader.py::test_slicer_writer_and_reader
tests/comoving/test_slicer_writer_and_reader.py::test_slicer_writer_and_reader
  /cosma/home/durham/dc-oman1/.virtualenv/paicos_review/lib/python3.10/site-packages/cupy/__init__.py:129: DeprecationWarning: `np.uint0` is a deprecated alias for `np.uintp`.  (Deprecated NumPy 1.24)
    from numpy import uint0  # NOQA

tests/comoving/test_slicer_writer_and_reader.py::test_slicer_writer_and_reader
tests/comoving/test_projector.py::test_projector
  /cosma/home/durham/dc-oman1/code/paicos/paicos/image_creators/image_creator.py:70: UserWarning:

  The definition of a 'y'-image was changed on 19/1/24 to be consistent with a right-handed coordinate system. You can recover the old behaviour by passing an orientation object:
  Orientation(normal_vector=[0, -1, 0], perp_vector1=[1, 0, 0])
  instead of a direction 'y' string.

    warnings.warn("\n\nThe definition of a 'y'-image was changed "

-- Docs: https://docs.pytest.org/en/stable/how-to/capture-warnings.html
======================== 9 passed, 8 warnings in 10.50s ========================
cd tests; pytest non-comoving
rootdir: /cosma/home/durham/dc-oman1/code/paicos
plugins: order-1.2.0
collected 1 item

non-comoving/test_non_comoving.py .                                      [100%]

=============================== warnings summary ===============================
tests/non-comoving/test_non_comoving.py::test_non_comoving
tests/non-comoving/test_non_comoving.py::test_non_comoving
  /cosma/home/durham/dc-oman1/.virtualenv/paicos_review/lib/python3.10/site-packages/cupy/__init__.py:103: DeprecationWarning: `np.bool8` is a deprecated alias for `np.bool_`.  (Deprecated NumPy 1.24)
    from numpy import bool8  # NOQA

tests/non-comoving/test_non_comoving.py::test_non_comoving
tests/non-comoving/test_non_comoving.py::test_non_comoving
  /cosma/home/durham/dc-oman1/.virtualenv/paicos_review/lib/python3.10/site-packages/cupy/__init__.py:114: DeprecationWarning: `np.int0` is a deprecated alias for `np.intp`.  (Deprecated NumPy 1.24)
    from numpy import int0  # NOQA

tests/non-comoving/test_non_comoving.py::test_non_comoving
tests/non-comoving/test_non_comoving.py::test_non_comoving
  /cosma/home/durham/dc-oman1/.virtualenv/paicos_review/lib/python3.10/site-packages/cupy/__init__.py:129: DeprecationWarning: `np.uint0` is a deprecated alias for `np.uintp`.  (Deprecated NumPy 1.24)
    from numpy import uint0  # NOQA

-- Docs: https://docs.pytest.org/en/stable/how-to/capture-warnings.html
======================== 1 passed, 6 warnings in 2.58s =========================
cd tests; pytest general
rootdir: /cosma/home/durham/dc-oman1/code/paicos
plugins: order-1.2.0
collected 1 item

general/test_orientation.py .                                            [100%]

=============================== warnings summary ===============================
../../../.virtualenv/paicos_review/lib/python3.10/site-packages/cupy/__init__.py:103
../../../.virtualenv/paicos_review/lib/python3.10/site-packages/cupy/__init__.py:103
  /cosma/home/durham/dc-oman1/.virtualenv/paicos_review/lib/python3.10/site-packages/cupy/__init__.py:103: DeprecationWarning: `np.bool8` is a deprecated alias for `np.bool_`.  (Deprecated NumPy 1.24)
    from numpy import bool8  # NOQA

../../../.virtualenv/paicos_review/lib/python3.10/site-packages/cupy/__init__.py:114
../../../.virtualenv/paicos_review/lib/python3.10/site-packages/cupy/__init__.py:114
  /cosma/home/durham/dc-oman1/.virtualenv/paicos_review/lib/python3.10/site-packages/cupy/__init__.py:114: DeprecationWarning: `np.int0` is a deprecated alias for `np.intp`.  (Deprecated NumPy 1.24)
    from numpy import int0  # NOQA

../../../.virtualenv/paicos_review/lib/python3.10/site-packages/cupy/__init__.py:129
../../../.virtualenv/paicos_review/lib/python3.10/site-packages/cupy/__init__.py:129
  /cosma/home/durham/dc-oman1/.virtualenv/paicos_review/lib/python3.10/site-packages/cupy/__init__.py:129: DeprecationWarning: `np.uint0` is a deprecated alias for `np.uintp`.  (Deprecated NumPy 1.24)
    from numpy import uint0  # NOQA

-- Docs: https://docs.pytest.org/en/stable/how-to/capture-warnings.html
======================== 1 passed, 6 warnings in 2.17s =========================
```
