### 2024-02-20 tberlok/paicos#48 @kyleaoman
There is also an implicit dependency on `ipympl`.
