### 2024-02-20 tberlok/paicos#56 @kyleaoman
Interactive demo of GPU rendering error
I've attempted to try out the example notebook `gpu_ray_with_widget_and_logging_v2.ipynb`. The other example notebooks showcasing the GPU functionality run fine for me on the same environment so it seems like I have the GPU configured correctly. My suspicion is that this might be a jupyterlab compatibility issue. Some details:

1. I'm running on the cosma system, specifically node login8b, which is equipped with an nvidia A100 GPU (https://www.dur.ac.uk/icc/cosma/support/gpu/).
2. I'm using cuda version 11.2, which the HPC admin team has installed as a module. I'm not entirely confident about loading a module for use inside a jupyterlab environment so I queried the module (which is just a set of instructions for setting environment variables) and added a cell to the top of the notebook to set those variables through `os.environ`.
3. I'm working in a virtual environment with the following packages installed (everything done with pip, I used the third method to install paicos, i.e. clone repository and then install with pip):
```
> pip freeze
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astropy==6.0.0
astropy-iers-data==0.2024.2.19.0.28.47
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bleach==6.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
comm==0.2.1
contourpy==1.2.0
cupy-cuda112==10.6.0
cycler==0.12.1
Cython==3.0.8
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
fastrlock==0.8.2
fonttools==4.49.0
fqdn==1.5.1
h11==0.14.0
h5py==3.10.0
httpcore==1.0.3
httpx==0.26.0
idna==3.6
iniconfig==2.0.0
ipykernel==6.29.2
ipympl==0.9.3
ipython==8.21.0
ipython-genutils==0.2.0
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
json5==0.9.17
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.1.2
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.3
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
llvmlite==0.42.0
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mistune==3.0.2
nbclient==0.9.0
nbconvert==7.16.1
nbformat==5.9.2
nest-asyncio==1.6.0
notebook==7.1.0
notebook_shim==0.2.4
numba==0.59.0
numpy==1.24.4
overrides==7.7.0
packaging==23.2
-e git+ssh://git@github.com/tberlok/paicos.git@4cca5893062d5435a1e6d849e3113d47e62dbd8a#egg=paicos
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
pluggy==1.4.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pyerfa==2.0.1.1
Pygments==2.17.2
pyparsing==3.1.1
pytest==8.0.1
pytest-order==1.2.0
python-dateutil==2.8.2
python-json-logger==2.0.7
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.33.0
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.18.0
scipy==1.12.0
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
terminado==0.18.0
tinycss2==1.2.1
tomli==2.0.1
tornado==6.4
traitlets==5.14.1
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
uri-template==1.3.0
urllib3==2.2.1
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
widgetsnbextension==4.0.10
```
4. I'm working in a jupyterlab instance hosted on the cosma jupyterhub (https://www.dur.ac.uk/icc/cosma/support/jupyterhub/). I have a suspicion that this may be the root of the problem since jupyter-notebook and jupyterlab have some compatibility issues. Might be possible to test on the TNG jupyterlab instance, but I don't know if they have any GPUs available and if so, how to use them...
5. The notebook runs fine without errors, but instead of an interactive interface at the end of the last cell I just get 11 copies of the message `Error displaying widget: model not found`

I can try to talk to our system admins in the coming days to see if there's a good way of starting a notebook server to use the GPUs on our system (rather than a jupyterhub/jupyterlab server) - a priori I don't know a way to do this on this system that doesn't involve an agonizingly slow remotely-hosted browser session.
