### 2024-02-20 #6296 @kyleaoman
Alright added one last issue around the GPU interactive features. My review is complete, so I leave it with the authors to address the github issues that I've opened to their satisfaction. Happy to iterate on things as needed, just let me know if you (authors/editors) need anything from me in the meantime.
