### 2024-02-28 tberlok/paicos#57 @ttricco
Community contributions
I'm doing my JOSS review of Paicos and I can't seem to find any information about how other users can contribute.

It would be great to have explicit statements somewhere about how users:
- could submit code or documentation (do you accept PRs from anyone?)
- notify of bugs or other problems with Paicos (should they use the issue tracker?)
- ask for help using the software (slack? email? issue tracker? go away and leave me alone?)
