### 2024-02-29 7a356a7
Filter cupy deprecation warnings using .pytest.ini
See issue #55
