### 2024-02-29 tberlok/paicos#55 @tberlok
These warnings came from cupy. I am now filtering them out, see https://github.com/tberlok/paicos/commit/7a356a75524b5aed0c9939186396dbd29e2a90b1
