### 2024-02-29 0eba077
Skip GPU tests if cupy and numba not installed
See issue #50
