### 2024-02-29 tberlok/paicos#51 @tberlok
This marking of the ordering of tests requires the pytest-order package:

https://pypi.org/project/pytest-order/

Can you please check that you have it installed?

pip install pytest-order
