### 2024-02-29 b0c07b3
Add GNU GENERAL PUBLIC LICENSE
#54
