### 2024-02-29 tberlok/paicos#51 @kyleaoman
Ah I see it in the `requirements.txt`.

Perhaps consider specifying any optional requirements in `setup.py`, e.g. so that `pip install paicos[dev]` installs dependencies for the test suite, and then pointing to that in contribution guidelines?
