+++ paicos/cython/get_derived_variables.pyx
@@ -13,19 +13,28 @@ ctypedef fused real_t:
 def get_curvature(real_t[:, :] Bvec, real_t[:, :] Bgradient):
 
     """
-    This function returns the curvature, equivalent
-    to code:
+    This function computes the magnetic field line curvature.
 
-    Kvec  = np.dot(Bgrad_mat[i,],Bvec[i,]) / B[i]**2
-    Kvec -= np.dot(np.dot(Bgrad_mat[i,],Bvec[i,]),Bvec[i,]) * Bvec[i,] / B[i]**4
-    K2 = (Kvec**2).sum())
-    K = np.sqrt(K2)
+    Parameters:
+        Bvec (array N, 3): Magnetic field vector
+        Bgradient (array N, 3, 3): Magnetic field gradient tensor
+
+    Returns:
+        array N: The magnetic field line curvature.
     """
     cdef int ip, ii, jj, kk
     cdef int Np = Bvec.shape[0]
     cdef real_t B2, gradB_ij, gradB_kj, term1, term2
 
     cdef real_t[:] K2 = np.zeros(Np, dtype=np.float64)
+
+    """
+    The code is equivalent to
+    Kvec  = np.dot(Bgrad_mat[i,],Bvec[i,]) / B[i]**2
+    Kvec -= np.dot(np.dot(Bgrad_mat[i,],Bvec[i,]),Bvec[i,]) * Bvec[i,] / B[i]**4
+    K2 = (Kvec**2).sum())
+    K = np.sqrt(K2)
+    """
 
     # Bgradient has not been reshaped, \nabla B_ij
     for ip in prange(Np, nogil=True, schedule='static'): #, num_threads=32):
@@ -53,7 +62,7 @@ def get_curvature(real_t[:, :] Bvec, real_t[:, :] Bgradient):
 def get_magnitude_of_vector(real_t[:, :] Bvec):
 
     """
-    Compute the magnitude of a vector, e.g., magnetic field strength
+    Computes the magnitude of a vector, e.g., magnetic field strength
     """
     cdef int ip, ii
     cdef int Np = Bvec.shape[0]
@@ -76,8 +85,8 @@ def get_magnitude_of_vector(real_t[:, :] Bvec):
 def sum_1d_array_omp(real_t[:] arr, int num_threads):
 
     """
-    Compute sum of an array along the first index.
-    Equivalent to np.sum(arr, axis=0)
+    Computes the sum of a 1D array using openmp.
+    Equivalent to np.sum(arr)
     """
     cdef int ip
     cdef int Np = arr.shape[0]
@@ -91,7 +100,7 @@ def sum_1d_array_omp(real_t[:] arr, int num_threads):
 def sum_2d_array_omp(real_t[:, :] arr, int num_threads):
 
     """
-    Compute sum of an array along the first index.
+    Computes the sum of an array along the first index using openmp.
     Equivalent to np.sum(arr, axis=0)
     """
     cdef int ip
@@ -111,8 +120,8 @@ def sum_2d_array_omp(real_t[:, :] arr, int num_threads):
 def sum_arr_times_vector_omp(real_t[:] arr, real_t[:, :] vector, int num_threads):
 
     """
-    Compute the sum of an array times a vector.
-    e.g. sum_i M_i \vec{v}_i for calculating the center of mass.
+    Computes the sum of an array times a vector.
+    e.g. sum_i M_i vec{v}_i for calculating the center of mass.
 
     Equivalent to np.sum(arr[:, None] * vector, axis=0)
     """
