### 2024-03-04 ee25f9e
Radial selection plus thin layer
