+++ examples/user_defined_functions_example.py
@@ -4,14 +4,14 @@ pa.use_units(True)
 pa.use_only_user_functions(False)
 
 
-def TemperaturesTimesMassesSquared(snap, get_depencies=False):
-    if get_depencies:
+def TemperaturesTimesMassesSquared(snap, get_dependencies=False):
+    if get_dependencies:
         return ['0_Temperatures', '0_Masses']
     return snap['0_Temperatures'] * snap['0_Masses']**2
 
 
-def CosmicRaysTimesMasses(snap, get_depencies=False):
-    if get_depencies:
+def CosmicRaysTimesMasses(snap, get_dependencies=False):
+    if get_dependencies:
         return ['0_CosmicRays', '0_Masses']
     return snap['0_CosmicRays'] * snap['0_Masses']**2
 
