+++ docs/source/configuration.md
@@ -56,8 +56,8 @@ Below is an example where the user would like to have Paicos
 be abble to automatically compute a derived quantity:
 ```
 import paicos as pa
-def TemperaturesTimesMassesSquared(snap, get_depencies=False):
-    if get_depencies:
+def TemperaturesTimesMassesSquared(snap, get_dependencies=False):
+    if get_dependencies:
         return ['0_Temperatures', '0_Masses']
     return snap['0_Temperatures'] * snap['0_Masses']**2
 
