### 2024-03-04 500974c
Update arepo_catalog.py
