### 2024-03-04 be2a57d
Update arepo_catalog.py
