+++ paicos/units.py
@@ -53,6 +53,8 @@ def get_unit_dictionaries(unit):
     """
     Returns dictionaries with information about the units of the
     quantity.
+
+    :meta private:
     """
     codic = {}
     dic = {}
@@ -71,7 +73,9 @@ def get_unit_dictionaries(unit):
 def construct_unit_from_dic(dic):
     """
     Construct unit from a dictionary with the format returned
-    from __get_unit_dictionaries
+    from get_unit_dictionaries
+
+    :meta private:
     """
     return np.prod([unit**dic[unit] for unit in dic])
 
@@ -95,6 +99,8 @@ def separate_units(unit):
 def get_new_unit(unit, remove_list=[]):
     """
     Return new unit where base units in the remove list have been removed.
+
+    :meta private:
     """
     unit_list = []
     for base, power in zip(unit.bases, unit.powers):
@@ -122,42 +128,48 @@ class PaicosQuantity(Quantity):
 
     h: the reduced Hubble parameter, e.g. h = 0.7
 
-    unit: a string, e.g. 'g/cm^3 small_a^-3 small_h^2' or astropy Unit
-          The latter can be defined like this:
+    unit: a string, e.g. ``g/cm^3 small_a^-3 small_h^2`` or an astropy Unit
+          The latter can be defined like this::
 
-          from paicos import units as pu
-          from astropy import units as u
-          unit = u.g*u.cm**(-3)*small_a**(-3)*small_h**(2)
+            from paicos import units as pu
+            from astropy import units as u
+            unit = u.g*u.cm**(-3)*small_a**(-3)*small_h**(2)
 
           The naming of small_a and small_h is to avoid conflict with the already
           existing 'annum' (i.e. a year) and 'h' (hour) units.
 
 
-    Methods/properties
-    ----------
+    Methods/key properties
+    ----------------------
 
-    no_small_h: returns a new comoving quantity where the h-factors have
-               been removed and the numeric value adjusted accordingly.
+        no_small_h
+                    Returns a new comoving quantity where the h-factors
+                    have been removed and the numeric value adjusted
+                    accordingly.
 
-    to_physical: returns a new  object where both a and h factors have been
-                 removed, i.e. we have switched from comoving values to
-                 the physical value.
+        to_physical
+                      Returns a new  object where both a and h factors have been
+                      removed, i.e. we have switched from comoving values to
+                      the physical value.
 
-    label: Return a Latex label for use in plots.
+        label
+                Returns a Latex label for use in plots.
 
     Examples
-    ----------
+    --------
 
-    units = 'g cm^-3 small_a^-3 small_h^2'
-    A = PaicosQuantity(2, units, h=0.7, a=1/128)
+    Here is an example::
 
-    # Create a new comoving quantity where the h-factors have been removed
-    B = A.no_small_h
+        units = 'g cm^-3 small_a^-3 small_h^2'
+        A = PaicosQuantity(2, units, h=0.7, a=1/128)
 
-    # Create a new quantity where both a and h factor have been removed,
-    # i.e. we have switched from a comoving quantity to the physical value
+        # Create a new comoving quantity where the h-factors have been removed
+        B = A.no_small_h
 
-    C = A.to_physical
+        # Create a new quantity where both a and h factor have been removed,
+        # i.e. we have switched from a comoving quantity to the physical value
+
+        C = A.to_physical
 
     """
 
@@ -172,6 +184,8 @@ class PaicosQuantity(Quantity):
         h
         a
         comoving_sim
+
+        :meta private:
         """
 
         assert h is not None, 'Paicos quantity is missing a value for h'
@@ -193,7 +207,9 @@ class PaicosQuantity(Quantity):
 
[…]
