+++ paicos/image_creators/slicer.py
@@ -34,10 +34,12 @@ class Slicer(ImageCreator):
             e.g. widths=[width_x, width_y, width_z] where one of the widths
             is zero (e.g. width_x=0 if direction='x').
 
-        direction : str
+        direction : str, Orientation
             Direction of the slicing, e.g. 'x', 'y' or 'z'. For instance,
             setting direction to 'x' gives a slice in the yz plane with the
             constant x-value set to be x_c.
+            For a more general orientation, one can pass an Orientation
+            instance.
 
         npix : int, optional
             Number of pixels in the horizontal direction of the image,
@@ -176,12 +178,13 @@ class Slicer(ImageCreator):
 
         Parameters
         ----------
-        variable: a string or an array of shape (N, ) representing the gas variable to slice.
-                  For instance '0_Density' or snap['0_Density'].
+            variable: str, arr
+                The variable to slice. For instance '0_Density' or snap['0_Density'].
 
         Returns
         -------
-            A 2D array the sliced variable.
+            slice : 2d arr
+                A 2d array of the sliced variable.
         """
 
         # This calls _do_region_selection if resolution, Orientation,
