### 2024-03-08 tberlok/paicos#49 @tberlok
I have explained in some details why it fails here: https://github.com/tberlok/paicos/issues/59

The problem here is that pytest uses one single Python session to run all the tests
when you do `pytest` at the base directory (this is why I introduced the  `make checks` command).
I have updated the error message that you get when loading several snapshots with different unit systems,
so that the difficult to understand and expected astropy error message is explained.

You can check that all the tests work like this (one python session for each set of tests):

```
pytest tests/comoving
pytest tests/non-comoving
pytest tests/general
pytest tests/cuda-gpu
```
This should hopefully work.
