### 2024-03-08 4e2d34d
Update contribution guidelines
See #47 and #57
