+++ paicos/readers/arepo_catalog.py
@@ -53,26 +53,34 @@ class Catalog(PaicosReader):
             The age of the Universe (only for cosmological runs).
 
     """
-    def __init__(self, basedir='.', snapnum=None,
+    def __init__(self, basedir='.', snapnum=None, load_all=True,
                  to_physical=False, subfind_catalog=True, verbose=False):
         """
         Initializes the Catalog class.
 
-        Parameters:
+        Parameters
+        ----------
 
-            basedir (str): path of the directory containing the catalogs
-                           (e.g. the 'output' folder).
+            basedir : str
+                The path of the directory containing the catalogs
+                (e.g. the 'output' folder).
 
-            snapnum (int): snapshot number.
+            snapnum : int
+                The snapshot number.
 
-            to_physical (bool): whether to convert to physical unit upon
-                                loading the data, default is False.
+            load_all : bool
+                Whether to immediately load all fields or not.
 
-            subfind_catalog (bool): whether the simulation has subfind catalogs,
-                                    when False the code will look for FoF-catalogs
-                                    only.
+            to_physical : bool
+                whether to convert to physical unit upon loading the data.
+                Default is False.
 
-            verbose (bool): whether to print information, default is False.
+            subfind_catalog : bool
+                whether the simulation has subfind catalogs,
+                when False the code will look for FoF-catalogs only.
+
+            verbose : bool
+                whether to print information, default is False.
 
         """
 
