### 2024-03-09 tberlok/paicos#56 @tberlok
Hi Kyle,

I am honestly not sure how to debug these kind of things. Looking on the web, I've also noticed the frequent issues with widgets and jupyterlab. I am myself using jupyter notebooks in the browser. Here is my environment:
```
[berlok@astro01 ~]$ pip list
Package                       Version      Editable project location
----------------------------- ------------ ---------------------------------
aiohttp                       3.8.5
aiosignal                     1.3.1
anyio                         4.0.0
appdirs                       1.4.4
argon2-cffi                   23.1.0
argon2-cffi-bindings          21.2.0
arrow                         1.2.3
astroid                       3.0.1
astropy                       5.3.3
asttokens                     2.4.0
async-lru                     2.0.4
async-timeout                 4.0.3
attrs                         23.1.0
awkward                       2.5.0
awkward-cpp                   26
Babel                         2.12.1
backcall                      0.2.0
backports.functools-lru-cache 1.6.5
beautifulsoup4                4.12.2
bleach                        6.0.0
boltons                       23.0.0
Brotli                        1.1.0
cached-property               1.5.2
cachetools                    5.3.2
certifi                       2023.7.22
cffi                          1.15.1
charset-normalizer            3.2.0
click                         8.1.7
cloudpickle                   3.0.0
cmasher                       1.6.3
cmyt                          1.4.0
colorama                      0.4.6
colorspacious                 1.1.2
comm                          0.1.4
conda                         23.7.4
conda-libmamba-solver         23.7.0
conda-package-handling        2.2.0
conda_package_streaming       0.9.0
contourpy                     1.1.1
cryptography                  41.0.4
cubinlinker-cu11              0.3.0.post1
cuda-python                   11.8.3
cudf-cu11                     23.10.2
cufinufft                     1.3
cuml-cu11                     23.10.0
cupy-cuda112                  10.6.0
cupy-knn                      0.2.5        /lustre/hpc/astro/berlok/cupy-knn
cycler                        0.11.0
Cython                        3.0.4
dask                          2023.9.2
dask-cuda                     23.10.0
dask-cudf-cu11                23.10.2
debugpy                       1.8.0
decorator                     5.1.1
defusedxml                    0.7.1
dill                          0.3.7
distributed                   2023.9.2
e13tools                      0.9.6
entrypoints                   0.4
ewah-bool-utils               1.1.0
exceptiongroup                1.1.3
executing                     1.2.0
fastjsonschema                2.18.0
fastrlock                     0.8.2
finufft                       2.1.0
flake8                        6.1.0
fonttools                     4.42.1
fortls                        2.13.0
fortran-language-server       1.12.0
fqdn                          1.5.1
frozenlist                    1.4.0
fsspec                        2023.10.0
gmpy2                         2.1.2
h5py                          3.9.0
idna                          3.4
importlib-metadata            6.8.0
importlib-resources           6.1.0
iniconfig                     2.0.0
ipykernel                     6.25.2
ipympl                        0.9.3
ipython                       8.15.0
ipython-genutils              0.2.0
ipywidgets                    8.1.1
isoduration                   20.11.0
isort                         5.12.0
jedi                          0.19.0
Jinja2                        3.1.2
joblib                        1.3.2
json5                         0.9.14
jsonpatch                     1.33
jsonpointer                   2.4
jsonschema                    4.19.1
jsonschema-specifications     2023.7.1
jupyter                       1.0.0
jupyter_client                8.3.1
jupyter-console               6.6.3
jupyter_core                  5.3.1
jupyter-events                0.7.0
jupyter-lsp                   2.2.0
jupyter_server                2.7.3
jupyter_server_terminals      0.4.4
jupyterlab                    4.0.6
jupyterlab-pygments           0.2.2
jupyterlab_server             2.25.0
jupyterlab-widgets            3.0.9
kiwisolver                    1.4.5
libmambapy                    1.5.1
llvmlite                      0.40.1
locket                        1.0.0
loguru                        0.7.2
Mako                          1.3.0
mamba                         1.5.1
markdown-it-py                3.0.0
MarkupSafe                    2.1.3
matplotlib                    3.8.0
matplotlib-inline             0.1.6
mccabe                        0.7.0
mdurl                         0.1.2
mistune                       3.0.1
more-itertools                10.1.0
mpmath                        1.3.0
msgpack                       1.0.7
multidict                     6.0.4
munkres                       1.1.4
nbclient                      0.8.0
nbconvert                     7.8.0
nbformat                      5.9.2
nest-asyncio                  1.5.6
notebook                      7.0.4
notebook_shim                 0.2.3
numba                         0.57.1
numpy                         1.24.4
nvtx                          0.2.8
overrides                     7.4.0
packaging                     23.1
paicos                        0.1.8
pandas                        1.5.3
pandocfilters                 1.5.0
parso                         0.8.3
partd                         1.4.1
pexpect                       4.8.0
pickleshare                   0.7.5
Pillow                        10.0.1
pip                           23.2.1
pkgutil_resolve_name          1.3.10
platformdirs                  3.10.0
pluggy                        1.3.0
ply                           3.11
pooch                         1.7.0
prometheus-client             0.17.1
prompt-toolkit                3.0.39
protobuf                      4.25.1
psutil                        5.9.5
ptxcompiler-cu11              0.7.0.post1
ptyprocess                    0.7.0
pure-eval                     0.2.2
pyarrow                       12.0.1
pycodestyle                   2.11.1
pycosat                       0.6.4
pycparser                     2.21
pycuda                        2023.1
pyerfa                        2.0.0.3
pyflakes                      3.1.0
Pygments                      2.16.1
pylibraft-cu11                23.10.0
pylint                        3.0.2
pynvml                        11.4.1
pyOpenSSL                     23.2.0
pyparsing                     3.1.1
PyQt5                         5.15.9
PyQt5-sip                     12.12.2
PySocks                       1.7.1
pytest                        7.4.3
pytest-order                  1.1.0
python-dateutil               2.8.2
python-json-logger            2.0.7
pytools                       2023.1.1
pytz                          2023.3.post1
pyvista                       0.42.2
PyYAML                        6.0.1
pyzmq                         25.1.1
qtconsole                     5.4.4
QtPy                          2.4.0
raft-dask-cu11                23.10.0
referencing                   0.30.2
requests                      2.31.0
rfc3339-validator             0.1.4
rfc3986-validator             0.1.1
rich                          13.7.0
rmm-cu11                      23.10.0
rpds-py                       0.10.3
ruamel.yaml                   0.17.32
ruamel.yaml.clib              0.2.7
scikit-learn                  1.3.2
scipy                         1.11.2
scooby                        0.7.3
Send2Trash                    1.8.2
setuptools                    68.2.2
sip                           6.7.11
six                           1.16.0
sniffio                       1.3.0
sortedcontainers              2.4.0
soupsieve                     2.5
stack-data                    0.6.2
sympy                         1.12
tblib                         3.0.0
terminado                     0.17.1
threadpoolctl                 3.2.0
tinycss2                      1.2.1
toml                          0.10.2
tomli                         2.0.1
tomli_w                       1.0.0
tomlkit                       0.12.1
toolz                         0.12.0
tornado                       6.3.3
tqdm                          4.66.1
traitlets                     5.10.1
treelite                      3.9.1
treelite-runtime              3.9.1
typing_extensions             4.8.0
typing-utils                  0.1.0
ucx-py-cu11                   0.34.0
unicodedata2                  15.0.0
unyt                          2.9.5
uri-template                  1.3.0
urllib3                       2.0.5
vtk                           9.2.6
wcwidth                       0.2.6
webcolors                     1.13
webencodings                  0.5.1
websocket-client              1.6.3
wheel                         0.41.2
widgetsnbextension            4.0.9
wslink                        1.11.4
xvfbwrapper                   0.2.9
yarl                          1.9.2
yt                            4.2.2
zict                          3.0.0
zipp                          3.17.0
zstandard                     0.21.0
```

What I have done is that I have added information about the requirements at the top of the notebooks, see e.g. here,
https://paicos.readthedocs.io/en/latest/notebooks/notebook6_interactive_visualization_on_the_gpu.html

If you still can't get it working, then I am unsure about what the next steps are as it's outside my expertise to provide user support on ipywidgets.  I might consider keeping the example in the repo but removing the highlight of this capability from the JOSS paper.

The other thing I will do is to create a simple python script for the GPU timings, so that you can check the timings outside of the notebook environments.

Best,
Thomas
