### 2024-03-09 tberlok/paicos#56 @kyleaoman
I think that signposting in the notebook that a standalone notebook server is likely to work but jupyterlab might not is probably enough. I'll email my HPC support team and see if there's any way I can try using the GPUs with a standalone notebook server (without X-forwarding a browser session, tried that and it's painfully unresponsive).
