### 2024-03-09 99d3dfb
Update gpu_sph_projector.py
