### 2024-03-09 5b02bed
Revert "Update gpu_sph_projector.py"
This reverts commit 99d3dfbdb9ff28f5aeb34604ca107cb5bbd286ef.
