+++ paicos/image_creators/gpu_sph_projector.py
@@ -172,7 +172,7 @@ class GpuSphProjector(ImageCreator):
 
     def __init__(self, snap, center, widths, direction,
                  npix=512, parttype=0, nvol=8, threadsperblock=8,
-                 do_pre_selection=False, sph_kernel_radius=4):
+                 do_pre_selection=False):
         """
         Initialize the Projector class.
 
@@ -201,10 +201,6 @@ class GpuSphProjector(ImageCreator):
 
         nvol : int, optional
             Integer used to determine the smoothing length, by default 8
-
-        sph_kernel_radius : int
-            Sets the kernel radius in units of the grid distance on the grid.
-            For testing purposes. 3 or 4 are probably reasonable values.
 
         """
 
@@ -242,7 +238,7 @@ class GpuSphProjector(ImageCreator):
         # TODO: add split into GPU and CPU based on cell sizes here.
         # Perhaps make the CPU do everything above grid8.
 
-        self._sph_kernel_radius = sph_kernel_radius
+        self._sph_kernel_radius = 4
 
         # Call selection
         self.has_do_region_selection_been_called = False
