### 2024-03-09 tberlok/paicos#57 @tberlok
There is now updated information on the README.md page of the Github repo. https://github.com/tberlok/paicos
The updates were also done following instructions in #47.
