### 2024-03-10 1a8e432
Saving reduced versions snap and catalog
Now also works for multi file snapshots and similar functionality implemented for catalogs.
