### 2024-03-10 5b242de
Added GPU speed test script
and instructions for downloading a sample data set
