### 2024-03-10 tberlok/paicos#54 @tberlok
Thanks for pointing this out. I have added a GNU general public license.
