### 2024-03-10 tberlok/paicos#48 @tberlok
I have added matplotlib as a dependency and now give instructions about ipywidgets in the relevant notebook: https://paicos.readthedocs.io/en/latest/notebooks/notebook6_interactive_visualization_on_the_gpu.html
