### 2024-03-10 tberlok/paicos#47 @tberlok
I have updated the guidelines to include these details: https://github.com/tberlok/paicos/blob/main/.github/CONTRIBUTING.md
