### 2024-03-10 #6296 @tberlok
Hi @kyleaoman, @ttricco, @JBorrow

First of all many thanks to Kyle and Terrence for the detailed comments on Paicos! This was just the sort of feedback that I was hoping to get from submitting to JOSS.  I believe I have now addressed all the issues that you raised on [Github](https://github.com/tberlok/paicos/issues), but please let me know if you agree, so that I can close them.

A remaining issue is that you both wanted to test the GPU functionality. @kyleaoman ran into the problem that setting up a notebook + GPU environment is more difficult than simply running a Python script from terminal. I have for this reason created a Python script, https://github.com/tberlok/paicos/blob/main/examples/gpu_sph_speed_test_example.py, which I hope will be more straightforward to run. This also now includes instructions for downloading a large data set, without which you can't test the speed of the implementation (as the GPU becomes under-utilized for small data sets).

When I ran this script myself, I found that the GPU SPH code is no longer as fast as it used to be. Using the A100, it now takes around 0.75 seconds to deposit 93 mio. particles onto a 4096^2 grid. I think I will change the wording in the paper from "it takes 0.15 seconds to project 93 mio. particles onto an image plane with $2048^2$ pixels." to be "it takes less than a second to project 100 mio. particles onto an image plane with $4096^2$ pixels. I believe it's possible to make this faster and will probably try to optimise at some point, but for now I would like to go ahead with this version of the code.

Best,
Thomas
