### 2024-03-10 tberlok/paicos#56 @tberlok
Thanks, I have done as suggested and added a warning about jupyter lab: https://paicos.readthedocs.io/en/latest/notebooks/notebook6_interactive_visualization_on_the_gpu.html

I have also added a high-resolution data set, so that getting it to work should be a bit more fun.
