### 2024-03-13 9f0cbc7
Regression fix and new test
Bug in the scaling of tree_projector with additive=True introduced in
b46555f61fb19c86f2ffbd3d56f92f8a80f9f47e

Now fixed and I also introduced a new test that would have caught this regression. Thanks to @lenajlassi for pointing this out!
