+++ examples/tree_projector_example.py
@@ -21,46 +21,51 @@ M = projector.project_variable('0_Masses')
 V = projector.project_variable('0_Volume')
 rho = M / V
 
-plt.figure(1)
-plt.clf()
-fig, axes = plt.subplots(num=1, nrows=3, sharex=True)
+rel_rho = (np.abs(tree_rho - rho) / tree_rho)
+rel_M = (np.abs(tree_M - M) / tree_M)
 
-if True:
-    # Define new imshow_wo_units that automatically plots the values
-    imshow_wo_units = pa.util.remove_astro_units(axes[0].imshow)
-    imshow_wo_units(rho, extent=projector.extent,
-                    vmin=rho.min(), vmax=rho.max())
-    axes[0].set_ylabel(projector.extent.label('y'))
-    axes[0].set_title('Sph projection')
+for ii in range(1, 3):
 
-    imshow_wo_units = pa.util.remove_astro_units(axes[1].imshow)
-    imshow_wo_units(tree_rho.value, extent=tree_projector.extent.value,
-                    vmin=rho.min(), vmax=rho.max())
-    axes[1].set_ylabel(tree_projector.extent.label('y'))
-    axes[1].set_title('Tree projection')
+    plt.figure(ii)
+    plt.clf()
+    fig, axes = plt.subplots(num=ii, nrows=3, sharex=True)
 
-    axes[2].imshow(np.abs(tree_rho.value - rho.value) / tree_rho.value,
-                   extent=tree_projector.extent.value)
-    axes[2].set_ylabel(tree_projector.extent.label('y'))
-    axes[2].set_xlabel(tree_projector.extent.label('x'))
-    axes[2].set_title('Rel difference')
+    if ii == 1:
+        # Define new imshow_wo_units that automatically plots the values
+        fig.suptitle('Density plots (by dividing additive projections)')
+        imshow_wo_units = pa.util.remove_astro_units(axes[0].imshow)
+        imshow_wo_units(rho, extent=projector.extent,
+                        vmin=rho.min(), vmax=rho.max())
+        axes[0].set_ylabel(projector.extent.label('y'))
+        axes[0].set_title('Sph projection')
 
-else:
-    # Define new imshow_wo_units that automatically plots the values
-    imshow_wo_units = pa.util.remove_astro_units(axes[0].imshow)
-    imshow_wo_units(M, extent=projector.extent)
-    axes[0].set_ylabel(projector.extent.label('y'))
-    axes[0].set_title('Sph projection')
+        imshow_wo_units = pa.util.remove_astro_units(axes[1].imshow)
+        imshow_wo_units(tree_rho.value, extent=tree_projector.extent.value,
+                        vmin=rho.min(), vmax=rho.max())
+        axes[1].set_ylabel(tree_projector.extent.label('y'))
+        axes[1].set_title('Tree projection')
 
-    imshow_wo_units = pa.util.remove_astro_units(axes[1].imshow)
-    imshow_wo_units(tree_M.value, extent=tree_projector.extent.value)
-    axes[1].set_ylabel(tree_projector.extent.label('y'))
-    axes[1].set_title('Tree projection')
+        axes[2].imshow(rel_rho.value, extent=tree_projector.extent.value)
+        axes[2].set_ylabel(tree_projector.extent.label('y'))
+        axes[2].set_xlabel(tree_projector.extent.label('x'))
+        axes[2].set_title('Rel difference')
 
-    axes[2].imshow(np.abs(tree_M.value - M.value) / tree_M.value,
-                   extent=tree_projector.extent.value)
-    axes[2].set_ylabel(tree_projector.extent.label('y'))
-    axes[2].set_xlabel(tree_projector.extent.label('x'))
[…]
