### 2024-03-15 c8b087e
Added GenericSnapshot
for experimental support for other codes
