### 2024-03-15 f3ae04d
changed cubic and spherical region selection
