+++ examples/example_conversions.py
@@ -9,4 +9,6 @@ snap['0_Temperatures']
 snap['0_Volume'].no_small_h.to('kpc3')
 snap['0_Volume'].to('kpc3').no_small_h
 
-snap['0_Volume'].to('kpc3 small_a^3 small_h^-2')
+snap['0_Volume'].to_any('kpc3 small_a^3 small_h^-2')
+
+## 
