### 2024-04-02 #6296 @kyleaoman
Hi all,

I finally managed to try out the interactive GPU visualization widget, works fine in a jupyter notebook (not lab). I'm satisfied with the changes to the package at this point and have ticked off the rest of my checklist.

@tberlok one remaining comment that I don't think should block publication. Installation is still a bit of a tricky process. Today I discovered that it seems like gcc (including recent versions) complains  about some of the warning flags and then exits, eventually got it to work with an intel compiler.
