### 2024-04-02 50460ca
Update GPU speed statement
