### 2024-04-05 #6296 @tberlok
Hi @JBorrow , are we sure that @ttricco has Github notifications enabled? Perhaps you can send him a direct message to make sure? Thanks.
