### 2024-04-09 50f542a
Make it possible to load GPU on demand
