+++ docs/requirements.txt
@@ -6,8 +6,6 @@
 #
 alabaster==0.7.13
     # via sphinx
-appnope==0.1.3
-    # via ipython
 astropy==5.3.4
     # via -r requirements.in
 asttokens==2.4.1
@@ -52,7 +50,7 @@ executing==2.0.1
     # via stack-data
 fastjsonschema==2.18.1
     # via nbformat
-h5py
+h5py==3.10.0
     # via -r requirements.in
 idna==3.4
     # via requests
@@ -62,7 +60,7 @@ ipython==8.16.1
     # via -r requirements.in
 jedi==0.19.1
     # via ipython
-jinja2==3.1.2
+jinja2==3.1.3
     # via
     #   myst-parser
     #   nbconvert
