+++ paicos/__init__.py
@@ -313,4 +313,3 @@ if settings.load_cuda_functionality_on_startup:
 
 # Do this at start up
 util._check_if_omp_has_issues()
-
