+++ setup.py
@@ -88,7 +88,7 @@ with open(cur + 'dev_requirements.txt') as f:
 
 setup(
     name='paicos',
-    version='0.1.8',
+    version='0.1.9',
     description=('An object-oriented Python package for analysis of '
                  + '(cosmological) simulations performed with Arepo.'),
     url='https://github.com/tberlok/paicos',
