+++ .github/workflows/manylinux.yml
@@ -16,11 +16,11 @@ jobs:
   deploy:
     runs-on: ubuntu-latest
     steps:
-    - uses: actions/checkout@v2
+    - uses: actions/checkout@v4
     - name: Set up Python
-      uses: actions/setup-python@v1
+      uses: actions/setup-python@v2
       with:
-        python-version: 3.10.13
+        python-version: 3.10.14
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
