+++ examples/example_conversions.py
@@ -11,4 +11,4 @@ snap['0_Volume'].to('kpc3').no_small_h
 
 snap['0_Volume'].to_any('kpc3 small_a^3 small_h^-2')
 
-## 
+##
