### 2024-04-11 08c25ec
User settings as hidden file in user directory
This is a much better and more standard place to put it for pip installations than the current location (which I think might be prone to accidental deletion).
