### 2024-04-11 6730634
Include image for display on pypi
