+++ MANIFEST.in
@@ -2,3 +2,4 @@ global-include *.pyx
 global-exclude *.c
 include requirements.txt
 include dev_requirements.txt
+include images/Z24_snap130_wide_projection_notnested.jpg
