### 2024-04-15 70471cf
Update requirements and bump version number
Some requirements now moved to dev_requirements
