+++ .github/workflows/pylint.yml
@@ -35,7 +35,7 @@ jobs:
         run: |
           python -m pip install --upgrade pip
           pip install -r requirements.txt
-          pip install pylint
+          pip install -r dev_requirements.txt
 
       - name: Install paicos
         run: make
