+++ .circleci/config.yml
@@ -35,6 +35,7 @@ jobs:
             python3 -m venv venv
             . venv/bin/activate
             pip install -r requirements.txt
+            pip install -r dev_requirements.txt
             pip install pytest-cov
             make
 
