### 2024-04-16 #6296 @tberlok
Hi @dfm, sorry for pinging you but it appears that the review of my paper has gone stale. That is, I have not heard from @ttricco since Feb 28 and it's been three weeks since @JBorrow sent the last reminder. As I understand from the JOSS guidelines, editors will normally ask reviewers to complete their review in 4-6 weeks and check in on this 1-2 times per week. So I am worried about what is going on and if my reminders above have been read. It will also soon be three months since submission (which I understand is the target duration for a JOSS review) and I am finding the silence and uncertainty about time scales difficult. Can you help me get this back on track and/or assure me that my submission has not been forgotten? Thanks!
