### 2024-04-18 #6296 @ttricco
@JBorrow I've managed this week to circle back around to finish my review. I've finished testing the software on two different linux environments / hardware and am satisfied that everything works as described.

Thanks for your patience @tberlok  and apologies for the delay on finalizing my review. The end of the teaching term was more overwhelming than I expected.
