### 2024-04-18 #6296 @tberlok
Wow, that went very fast all of a sudden! Thanks everyone, for reviewing and editing the code and paper!

I have created a Zenodo archive here: [10.5281/zenodo.10994256](https://zenodo.org/doi/10.5281/zenodo.10994256)
Here I have uploaded a tagged version of Paicos, with tag v0.1.14, https://github.com/tberlok/paicos/releases/tag/v0.1.14

I have checked the Zenodo metadata and added all our ORCIDs.
