### 2024-04-18 #6296 @tberlok
One last thing from my side today, I have just added "We thank the JOSS referees, Kyle Oman and Terrence Tricco, and the editor, Josh Borrow, for constructive feedback that greatly helped improve Paicos." to the paper.
