### 2024-04-19 003beac
Add conda-forge and PyPi details
