### 2024-04-19 ebb8d65
Update configuration.md
