### 2024-04-19 #6296 @tberlok
> One final comment on the paper: you have used `i.e.` on line 26, where I think you may mean `e.g.`.

@JBorrow Also thanks for pointing this out, I have made the correction. As I understand, this means that I have no more tasks left, but do let me know if you need anything else from me or if I need to make other final corrections.
