+++ paper/paper.bib
@@ -382,7 +382,7 @@ archivePrefix = {arXiv},
 @inproceedings{Kluyver2016,
        booktitle = {Positioning and Power in Academic Publishing: Players, Agents and Agendas},
           editor = {Fernando Loizides and Birgit Scmidt},
-           title = {Jupyter Notebooks ? a publishing format for reproducible computational workflows},
+           title = {Jupyter Notebooks -- a publishing format for reproducible computational workflows},
           author = {Thomas Kluyver and Benjamin Ragan-Kelley and Fernando P{\'e}rez and Brian Granger and Matthias Bussonnier and Jonathan Frederic and Kyle Kelley and Jessica Hamrick and Jason Grout and Sylvain Corlay and Paul Ivanov and Dami{\'a}n Avila and Safia Abdalla and Carol Willing and  Jupyter development team},
        publisher = {IOS Press},
             year = {2016},
