### 2024-04-19 #6296 @JBorrow
> There are a couple of missing DOIs in the references, though I have looked and it seems that indeed you follow their requested citation strategy.
>
> The only odd thing there @tberlok is the question mark in the Jupyter Notebooks title.

Additionally, that specific reference has a DOI that is not used here: http://dx.doi.org/10.3233/978-1-61499-649-1-87
