### 2024-04-19 31c05c4
Add jupyter DOI
