+++ paper/paper.bib
@@ -387,6 +387,7 @@ archivePrefix = {arXiv},
        publisher = {IOS Press},
             year = {2016},
            pages = {87--90},
+           doi = {10.3233/978-1-61499-649-1-87},
              url = {https://eprints.soton.ac.uk/403913/},
         abstract = {It is increasingly necessary for researchers in all fields to write computer code, and in order to reproduce research results, it is important that this code is published. We present Jupyter notebooks, a document format for publishing code, results and explanations in a form that is both readable and executable. We discuss various tools and use cases for notebook documents.}
 }
