### 2024-04-19 #6296 @tberlok
> There are a couple of missing DOIs in the references, though I have looked and it seems that indeed you follow their requested citation strategy.
>
> The only odd thing there @tberlok is the question mark in the Jupyter Notebooks title.

@JBorrow Thanks! I just now got rid of the question mark and added the DOI.
