### 2024-04-19 #6296 @tberlok
> That doesn't look like a valid DOI value

@JBorrow Not sure what is going here? I checked the zenodo DOI [here](https://dx.doi.org/) and it works.
