### 2024-04-19 #6296 @JBorrow
Seems like zenodo was just a little behind. One last check of the references and PDF, and I think we are good to go!
