### 2024-04-19 #6296 @JBorrow
Sorry @tberlok, would you please be able to add a DOI for the Maximizing Parallelism...

It's available here: http://diglib.eg.org/handle/10.2312/EGGH.HPG12.033-037
