+++ paper/paper.bib
@@ -128,6 +128,7 @@ address = {Goslar, DEU},
 abstract = {A number of methods for constructing bounding volume hierarchies and point-based octrees on the GPU are based on the idea of ordering primitives along a space-filling curve. A major shortcoming with these methods is that they construct levels of the tree sequentially, which limits the amount of parallelism that they can achieve. We present a novel approach that improves scalability by constructing the entire tree in parallel. Our main contribution is an in-place algorithm for constructing binary radix trees, which we use as a building block for other types of trees.},
 booktitle = {Proceedings of the Fourth ACM SIGGRAPH / Eurographics Conference on High-Performance Graphics},
 pages = {33–37},
+doi = {10.2312/EGGH/HPG12/033-037},
 numpages = {5},
 location = {Paris, France},
 series = {EGGH-HPG'12}
