### 2024-04-19 #6296 @tberlok
> Sorry @tberlok, would you please be able to add a DOI for the Maximizing Parallelism...
>
> It's available here: http://diglib.eg.org/handle/10.2312/EGGH.HPG12.033-037

@JBorrow Done - thanks for tracking these down!
