### 2024-04-20 e5e23d4
Edits for JOSS publication
