+++ paper/paper.bib
@@ -156,7 +156,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Somerville2015,
        author = {{Somerville}, Rachel S. and {Dav{\'e}}, Romeel},
         title = "{Physical Models of Galaxy Formation in a Cosmological Framework}",
-      journal = {\araa},
+      journal = {Annual Review of Astronomy and Astrophysics},
      keywords = {Astrophysics - Astrophysics of Galaxies},
          year = 2015,
         month = aug,
@@ -180,7 +180,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Springel2010,
        author = {{Springel}, Volker},
         title = "{E pur si muove: Galilean-invariant cosmological hydrodynamical simulations on a moving mesh}",
-      journal = {\mnras},
+      journal = {Monthly Notices of the RAS},
      keywords = {methods: numerical, galaxies: interactions, cosmology: dark matter, Astrophysics - Cosmology and Nongalactic Astrophysics},
          year = 2010,
         month = jan,
@@ -200,7 +200,7 @@ archivePrefix = {arXiv},
          {Skory}, Stephen and {Skillman}, Samuel W. and {Abel}, Tom and
          {Norman}, Michael L.},
         title = "{yt: A Multi-code Analysis Toolkit for Astrophysical Simulation Data}",
-      journal = {\apjs},
+      journal = {Astrophysical Journal, Supplement},
      keywords = {cosmology: theory, methods: data analysis, methods: numerical, Astrophysics - Instrumentation and Methods for Astrophysics, Astrophysics - Cosmology and Nongalactic Astrophysics},
          year = "2011",
         month = "Jan",
@@ -238,7 +238,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Harris2020,
        author = {{Harris}, Charles R. and {Millman}, K. Jarrod and {van der Walt}, St{\'e}fan J. and {Gommers}, Ralf and {Virtanen}, Pauli and {Cournapeau}, David and {Wieser}, Eric and {Taylor}, Julian and {Berg}, Sebastian and {Smith}, Nathaniel J. and {Kern}, Robert and {Picus}, Matti and {Hoyer}, Stephan and {van Kerkwijk}, Marten H. and {Brett}, Matthew and {Haldane}, Allan and {del R{\'\i}o}, Jaime Fern{\'a}ndez and {Wiebe}, Mark and {Peterson}, Pearu and {G{\'e}rard-Marchant}, Pierre and {Sheppard}, Kevin and {Reddy}, Tyler and {Weckesser}, Warren and {Abbasi}, Hameer and {Gohlke}, Christoph and {Oliphant}, Travis E.},
         title = "{Array programming with NumPy}",
-      journal = {\nat},
+      journal = {Nature},
      keywords = {Computer Science - Mathematical Software, Statistics - Computation},
          year = 2020,
         month = sep,
@@ -265,7 +265,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Pakmor2013,
        author = {{Pakmor}, R{\"u}diger and {Springel}, Volker},
         title = "{Simulations of magnetic fields in isolated disc galaxies}",
-      journal = {\mnras},
+      journal = {Monthly Notices of the RAS},
      keywords = {MHD, methods: numerical, galaxies: formation, Astrophysics - Cosmology and Extragalactic Astrophysics},
          year = 2013,
         month = jun,
@@ -283,7 +283,7 @@ archivePrefix = {arXiv},
 @ARTICLE{Springel2005,
[…]
